moved header and navside under /main
