--- src/app/header/index.ts
@@ -1,1 +1,0 @@
-export * from './header.module';
