+++ src/app/main/header/index.ts
@@ -1,0 +1,2 @@
+export * from './header.module';
+export * from './header.component';
