+++ src/app/main/index.ts
@@ -1,0 +1,2 @@
+export * from './main.module';
+export * from './main.page';
