+++ src/app/main/main.component.html
@@ -1,5 +1,5 @@
-<ds-header></ds-header>
 <router-outlet name="navSide"></router-outlet>
+<router-outlet name="header"></router-outlet>
 <div class="main-content">
 	<div class="container">
 		<router-outlet></router-outlet>
