+++ src/app/main/main.module.ts
@@ -1,12 +1,14 @@
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { MainComponent } from './main.component';
-import { DsHeaderModule } from '../header';
+import { DsHeaderModule } from './header';
+import { DsNavSideModule } from './nav-side';
 
 @NgModule({
 	imports: [
 		RouterModule,
-		DsHeaderModule
+		DsHeaderModule,
+		DsNavSideModule
 	],
 	declarations: [
 		MainComponent
