+++ src/app/main/nav-side/index.ts
@@ -1,0 +1,2 @@
+export * from './nav-side.module';
+export * from './nav-side.component';
