--- src/app/nav-side/index.ts
@@ -1,1 +1,0 @@
-export * from './nav-side.module';
